refactor(spid): Change `tpm_header_not_empty` to level
_Related Issue: https://github.com/lowRISC/opentitan/issues/15282_

The current interrupt scheme assumes the input is an event (pulse). If
the input is a status, it was recommented to add an edge detector to
create a pulse.

In certain cases, this may create unwanted (phantom) interrupts or
missing interrupts depending on the SW approach to handling the
interrupts.

To remove the race condition, HW now connects the status signal to the
interrupt in this commit. With the change, now SW is responsible to mask
the interrupt while the SW queues the interrupt to event queues.

Signed-off-by: Eli Kim <eli@opentitan.org>