[top-level] Fix csrng_lc_hw_debug_en_test
The lc_ctrl sets `lc_hw_debug` when the device is in test unlocked, dev
or rma mode. The test was previously expecting dev to behave as a prod
device by mistake. This change updates the test logic to expect
deterministic entropy to match if the lc_state is DEV.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>