[fpga] Add special case for OTP to gen_vivado_mem_image.py
This script reads MEM files that can be read by Verilog's $readmemh and
produces a MEM file that `updatemem` will understand. It turns out that
the transformation for ROM data was not right for OTP data due to
differences in the underlying hardware. This caught me by surprise
because updatemem was exiting successfully.

By comparing //hw/ip/otp_ctrl/data:img_rma, which is passed directly to
Vivado, with the INIT_XX strings that Vivado produces, I was able to
determine the transformation. This commit also includes a few features
that I used to bootstrap my understanding: (1) a function that reads
INIT_XX strings from OTP BRAMs and produces their memory content and (2)
a mocked out updatemem "simulator" that takes hex strings as input and
produces the same INIT_XX lines that the real updatemem would.

I used a modified version of the sram_program GDB test[0] to dump
relevant sections of OTP memory and verify that the OTP values I
modified can actually be read from OTP memory.

[0]: //sw/device/examples/sram_program:sram_program_fpga_cw310

Signed-off-by: Dan McArdle <dmcardle@google.com>