[test] Add chip_sw_sysrst_ctrl_in_irq
Test for SYSRST CTRL to detect input and key combo change.
On DV side the inputs of the SYSRST_CTRL are changed and
this change causes an interrupt on PLIC. Necessary
configuration are done for SYSRST_CTRL and PLIC. SW handles
the interrupt when triggerred, verifies the pin input value
and key_intr_status for correctness and clears the
interrupt status.

Signed-off-by: Abdullah Varici <abdullah.varici@lowrisc.org>