[utils,gpio] Add new hjson key 'auto_split' as hint for regtool
Introduces a new optional key 'auto_split' to serve as a hint
for register map generator script. Adds support up to 'auto_split'
in the TockOS register map generator.

If the bitfield or signal has key 'auto_split' set to value
true generator will split this filed to multiple one-bit fields.
The name of each generated bitfiled will consist of the
original bitfield name and an offset. If the parameter is set to false,
the current behavior of the regtool is not replaced. The default value of the
new key is set to false.

Signed-off-by: Michał Mazurek <maz@semihalf.com>