[dv/chip] Fix xcelium stub_cpu mode issue
The stub_cpu mode uses `CPU_HIER.clk_i` as the cpu_clk to drive tlul
transaction.
But in tb.sv, the current code force this wire to be 0 if we are in
stub_cpu mode.
I think the intention is to not force this clock in the stub_cpu mode.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>