[flash_ctrl] Handle faulted reads
In the current design, if a read were faulted to a non-existant location
AFTER the memory properties check, the return data could be unknown
and that makes the assertions go wild.

Since we are already detecting for this type of attack, use the same signals
to force the returning data and error conditions to a known value.

This removes the need for DV to individually turn off asertions for this
case.

Signed-off-by: Timothy Chen <timothytim@google.com>