[dv/top_earlgrey] Add test for uploaded flash commands
Add the option to the software to have write commands uploaded to
software instead of passed through. Use interrupts to trigger the
software respones, and have software copy the command out.

The virtual sequence checks that commands get uploaded as required and
that the SPI device agent receives all the appropriate commands, whether
they are copied out or passed through.

Signed-off-by: Alexander Williams <awill@google.com>