[spi_host dv] Fix drive conflict when spi_agent responsing read requests
Reorganize the fsm in decode_cmd. It was causing failure.
Set cfg.spi_mode for every transaction.
Set active_csb to 0. It is causing problems when it is randomly
chosen as we have only one agent connected to spi_host.
Increase the rx data size in order to prevent disabling
iso_fork until csb is deasserted.

Signed-off-by: Abdullah Varici <abdullah.varici@lowrisc.org>