[rom_e2e] fix (and simplify) sigverify_always test
This refactors the sigverify_always ROM E2E test to:
1. correct its implementation, adn
2. make it easier to read.

Regarding the first point, this test was incorrectly using the
two "empty_test" images to create a scenario where "no slots are
filled". The "empty_test_*" images are not empty in the sense that they
are all ones, rather they are a test that returns true.

After refactoring, the test now tests the following scenarios, as
requested
[here](https://github.com/lowRISC/opentitan/pull/15553#issuecomment-1287774364)
1. nothing in slot A (all ones); a corrupted image in slot B
2. nothing in slot B (all ones); a corrupted image in slot A
3. nothing in slot A (all ones); nothing in slot B (all ones)

Signed-off-by: Timothy Trippel <ttrippel@google.com>